requirements.txt
python-telegram-bot[all]>=20.0
httpx
python-dotenv
